feat: implementation

+++ src/main/java/jp/openstandia/connector/auth0/Auth0Connector.java
@@ -15,19 +15,9 @@
  */
 package jp.openstandia.connector.auth0;
 
-import com.auth0.client.HttpOptions;
-import com.auth0.client.ProxyOptions;
-import com.auth0.client.auth.AuthAPI;
-import com.auth0.client.mgmt.ManagementAPI;
-import com.auth0.client.mgmt.filter.UserFilter;
 import com.auth0.exception.APIException;
 import com.auth0.exception.Auth0Exception;
 import com.auth0.exception.RateLimitException;
-import com.auth0.json.auth.TokenHolder;
-import com.auth0.json.mgmt.users.UsersPage;
-import com.auth0.net.AuthRequest;
-import com.auth0.net.Request;
-import org.identityconnectors.common.StringUtil;
 import org.identityconnectors.common.logging.Log;
 import org.identityconnectors.framework.common.exceptions.*;
 import org.identityconnectors.framework.common.objects.*;
@@ -37,28 +27,28 @@ import org.identityconnectors.framework.spi.ConnectorClass;
 import org.identityconnectors.framework.spi.InstanceNameAware;
 import org.identityconnectors.framework.spi.PoolableConnector;
 import org.identityconnectors.framework.spi.operations.*;
-import software.amazon.awssdk.services.cognitoidentityprovider.CognitoIdentityProviderClient;
-import software.amazon.awssdk.services.cognitoidentityprovider.model.ResourceNotFoundException;
-import software.amazon.awssdk.services.cognitoidentityprovider.model.UserNotFoundException;
 
-import java.net.InetSocketAddress;
-import java.net.Proxy;
-import java.util.*;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
 
+import static jp.openstandia.connector.auth0.Auth0OrganizationHandler.ORGANIZATION_OBJECT_CLASS;
 import static jp.openstandia.connector.auth0.Auth0RoleHandler.ROLE_OBJECT_CLASS;
 import static jp.openstandia.connector.auth0.Auth0UserHandler.USER_OBJECT_CLASS;
 
-@ConnectorClass(configurationClass = Auth0Configuration.class, displayNameKey = "NRI OpenStandia Amazon Cognito User Pool Connector")
+@ConnectorClass(configurationClass = Auth0Configuration.class, displayNameKey = "Auth0 Connector")
 public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaOp, DeleteOp, SchemaOp, TestOp, SearchOp<Auth0Filter>, InstanceNameAware {
 
     private static final Log LOG = Log.getLog(Auth0Connector.class);
 
     protected Auth0Configuration configuration;
-    protected CognitoIdentityProviderClient client;
-    protected ManagementAPI client2;
-    protected TokenHolder tokenHolder;
+    protected Auth0Client client;
 
     private Map<String, AttributeInfo> userSchemaMap;
+    private Map<String, AttributeInfo> roleSchemaMap;
+    private Map<String, AttributeInfo> organizationSchemaMap;
+
     private String instanceName;
 
     @Override
@@ -69,89 +59,15 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
     @Override
     public void init(Configuration configuration) {
         this.configuration = (Auth0Configuration) configuration;
-
-        try {
-            initClient();
-        } catch (Exception e) {
-            throw processException(e);
-        }
-
+        initClient();
         LOG.ok("Connector {0} successfully initialized", getClass().getName());
     }
 
-    protected void initClient() throws Auth0Exception {
-        HttpOptions httpOptions = new HttpOptions();
-
-        if (configuration.getConnectionTimeoutInSeconds() != null) {
-            httpOptions.setConnectTimeout(configuration.getConnectionTimeoutInSeconds());
-        }
-        if (configuration.getReadTimeoutInSeconds() != null) {
-            httpOptions.setReadTimeout(configuration.getReadTimeoutInSeconds());
-        }
-        if (configuration.getMaxRetries() != null) {
-            httpOptions.setManagementAPIMaxRetries(configuration.getMaxRetries());
-        }
-
-        // HTTP Proxy
-        if (StringUtil.isNotEmpty(configuration.getHttpProxyHost())) {
-            Proxy proxy = new Proxy(Proxy.Type.HTTP, new InetSocketAddress(configuration.getHttpProxyHost(), configuration.getHttpProxyPort()));
-            ProxyOptions proxyOptions = new ProxyOptions(proxy);
-            httpOptions.setProxyOptions(proxyOptions);
-        }
-
-        // Setup client
-        refreshToken();
-        client2 = new ManagementAPI(configuration.getDomain(), tokenHolder.getAccessToken(), httpOptions);
-
-        // Verify we can access the API
-        checkClient();
-    }
-
-    protected void refreshToken() throws Auth0Exception {
-        if (configuration.getClientId() != null && configuration.getClientSecret() != null) {
-            if (isExpired(tokenHolder)) {
-                final AuthAPI[] authAPI = new AuthAPI[1];
-                configuration.getClientSecret().access(c -> {
-                    authAPI[0] = new AuthAPI(configuration.getDomain(), configuration.getDomain(), String.valueOf(c));
-                });
-                AuthRequest authRequest = authAPI[0].requestToken(String.format("https://%s/api/v2/", configuration.getDomain()));
-                tokenHolder = authRequest.execute();
-
-                if (client2 != null) {
-                    client2.setApiToken(tokenHolder.getAccessToken());
-                }
-            }
-        } else {
-            throw new ConfigurationException("Not configured Client ID or Client Secret for the API client");
-        }
-    }
-
-    protected boolean isExpired(TokenHolder holder) {
-        if (holder == null) {
-            return true;
-        }
-        long expiresAt = holder.getExpiresAt().getTime();
-        long now = new Date().getTime();
-
-        if (expiresAt + (60 * 1000) > now) {
-            LOG.ok("Detected the token is expired");
-            return true;
-        }
-
-        return false;
-    }
-
-    private void checkClient() {
-        if (client2 == null) {
-            throw new ConfigurationException("Not initialized the API client");
-        }
-        UserFilter filter = new UserFilter()
-                .withPage(0, 1);
-        Request<UsersPage> request = client2.users().list(filter);
-
+    protected void initClient() {
+        this.client = new Auth0Client();
         try {
-            UsersPage response = request.execute();
-        } catch (Auth0Exception e) {
+            client.initClient(configuration);
+        } catch (Exception e) {
             throw processException(e);
         }
     }
@@ -161,11 +77,14 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
         try {
             SchemaBuilder schemaBuilder = new SchemaBuilder(Auth0Connector.class);
 
-            ObjectClassInfo userSchemaInfo = Auth0UserHandler.getUserSchema(configuration);
+            ObjectClassInfo userSchemaInfo = Auth0UserHandler.getSchema(configuration);
             schemaBuilder.defineObjectClass(userSchemaInfo);
 
-            ObjectClassInfo groupSchemaInfo = Auth0RoleHandler.getGroupSchema(configuration);
-            schemaBuilder.defineObjectClass(groupSchemaInfo);
+            ObjectClassInfo roleSchemaInfo = Auth0RoleHandler.getSchema(configuration);
+            schemaBuilder.defineObjectClass(roleSchemaInfo);
+
+            ObjectClassInfo organizationSchemaInfo = Auth0OrganizationHandler.getSchema(configuration);
+            schemaBuilder.defineObjectClass(organizationSchemaInfo);
 
             schemaBuilder.defineOperationOption(OperationOptionInfoBuilder.buildAttributesToGet(), SearchOp.class);
             schemaBuilder.defineOperationOption(OperationOptionInfoBuilder.buildReturnDefaultAttributes(), SearchOp.class);
@@ -174,6 +93,16 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
             userSchemaInfo.getAttributeInfo().stream()
                     .forEach(a -> userSchemaMap.put(a.getName(), a));
             userSchemaMap = Collections.unmodifiableMap(userSchemaMap);
+
+            roleSchemaMap = new HashMap<>();
+            roleSchemaInfo.getAttributeInfo().stream()
+                    .forEach(a -> roleSchemaMap.put(a.getName(), a));
+            roleSchemaMap = Collections.unmodifiableMap(roleSchemaMap);
+
+            organizationSchemaMap = new HashMap<>();
+            organizationSchemaInfo.getAttributeInfo().stream()
+                    .forEach(a -> organizationSchemaMap.put(a.getName(), a));
+            organizationSchemaMap = Collections.unmodifiableMap(organizationSchemaMap);
 
             return schemaBuilder.build();
 
@@ -190,6 +119,22 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
         return userSchemaMap;
     }
 
+    private Map<String, AttributeInfo> getRoleSchemaMap() {
+        // Load schema map if it's not loaded yet
+        if (roleSchemaMap == null) {
+            schema();
+        }
+        return roleSchemaMap;
+    }
+
+    private Map<String, AttributeInfo> getOrganizationSchemaMap() {
+        // Load schema map if it's not loaded yet
+        if (organizationSchemaMap == null) {
+            schema();
+        }
+        return organizationSchemaMap;
+    }
+
     @Override
     public Uid create(ObjectClass objectClass, Set<Attribute> createAttributes, OperationOptions options) {
         if (objectClass == null) {
@@ -197,18 +142,22 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
         }
         LOG.info("CREATE METHOD OBJECTCLASS VALUE: {0}", objectClass);
 
-        if (createAttributes == null) {
-            throw new InvalidAttributeValueException("Attributes not provided or empty");
+        if (createAttributes == null || createAttributes.isEmpty()) {
+            throw new InvalidAttributeValueException("attributes not provided or empty");
         }
 
         try {
             if (objectClass.equals(USER_OBJECT_CLASS)) {
-                Auth0UserHandler userHandler = new Auth0UserHandler(configuration, client2, getUserSchemaMap());
+                Auth0UserHandler userHandler = new Auth0UserHandler(configuration, client, getUserSchemaMap());
                 return userHandler.createUser(createAttributes);
 
             } else if (objectClass.equals(ROLE_OBJECT_CLASS)) {
-                Auth0RoleHandler roleHandler = new Auth0RoleHandler(configuration, client);
+                Auth0RoleHandler roleHandler = new Auth0RoleHandler(configuration, client, getRoleSchemaMap());
                 return roleHandler.createRole(createAttributes);
+
+            } else if (objectClass.equals(ORGANIZATION_OBJECT_CLASS)) {
+                Auth0OrganizationHandler organizationHandler = new Auth0OrganizationHandler(configuration, client, getOrganizationSchemaMap());
+                return organizationHandler.create(createAttributes);
 
             } else {
                 throw new InvalidAttributeValueException("Unsupported object class " + objectClass);
@@ -226,8 +175,12 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
                 return userHandler.updateDelta(uid, modifications, options);
 
             } else if (objectClass.equals(ROLE_OBJECT_CLASS)) {
-                Auth0RoleHandler roleHandler = new Auth0RoleHandler(configuration, client);
+                Auth0RoleHandler roleHandler = new Auth0RoleHandler(configuration, client, getRoleSchemaMap());
                 return roleHandler.updateDelta(uid, modifications, options);
+
+            } else if (objectClass.equals(ORGANIZATION_OBJECT_CLASS)) {
+                Auth0OrganizationHandler organizationHandler = new Auth0OrganizationHandler(configuration, client, getOrganizationSchemaMap());
+                return organizationHandler.updateDelta(uid, modifications, options);
 
             } else {
                 throw new InvalidAttributeValueException("Unsupported object class " + objectClass);
@@ -239,14 +192,22 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
 
     @Override
     public void delete(ObjectClass objectClass, Uid uid, OperationOptions options) {
+        if (uid == null) {
+            throw new InvalidAttributeValueException("uid not provided");
+        }
+
         try {
             if (objectClass.equals(USER_OBJECT_CLASS)) {
                 Auth0UserHandler userHandler = new Auth0UserHandler(configuration, client, getUserSchemaMap());
                 userHandler.deleteUser(uid, options);
 
             } else if (objectClass.equals(ROLE_OBJECT_CLASS)) {
-                Auth0RoleHandler roleHandler = new Auth0RoleHandler(configuration, client);
-                roleHandler.deleteRole(objectClass, uid, options);
+                Auth0RoleHandler roleHandler = new Auth0RoleHandler(configuration, client, getRoleSchemaMap());
+                roleHandler.deleteRole(uid, options);
+
+            } else if (objectClass.equals(ORGANIZATION_OBJECT_CLASS)) {
+                Auth0OrganizationHandler organizationHandler = new Auth0OrganizationHandler(configuration, client, getOrganizationSchemaMap());
+                organizationHandler.delete(uid, options);
 
             } else {
                 throw new InvalidAttributeValueException("Unsupported object class " + objectClass);
@@ -263,30 +224,24 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
 
     @Override
     public void executeQuery(ObjectClass objectClass, Auth0Filter filter, ResultsHandler resultsHandler, OperationOptions options) {
-        if (objectClass.equals(USER_OBJECT_CLASS)) {
-            try {
+        try {
+            if (objectClass.equals(USER_OBJECT_CLASS)) {
                 Auth0UserHandler userHandler = new Auth0UserHandler(configuration, client, getUserSchemaMap());
                 userHandler.getUsers(filter, resultsHandler, options);
-            } catch (UserNotFoundException e) {
-                // Don't throw UnknownUidException
-                return;
-            } catch (Exception e) {
-                throw processException(e);
-            }
 
-        } else if (objectClass.equals(ROLE_OBJECT_CLASS)) {
-            try {
-                Auth0RoleHandler roleHandler = new Auth0RoleHandler(configuration, client);
+            } else if (objectClass.equals(ROLE_OBJECT_CLASS)) {
+                Auth0RoleHandler roleHandler = new Auth0RoleHandler(configuration, client, getRoleSchemaMap());
                 roleHandler.getRoles(filter, resultsHandler, options);
-            } catch (ResourceNotFoundException e) {
-                // Don't throw UnknownUidException
-                return;
-            } catch (Exception e) {
-                throw processException(e);
-            }
 
-        } else {
-            throw new InvalidAttributeValueException("Unsupported object class " + objectClass);
+            } else if (objectClass.equals(ORGANIZATION_OBJECT_CLASS)) {
+                Auth0OrganizationHandler organizationHandler = new Auth0OrganizationHandler(configuration, client, getOrganizationSchemaMap());
+                organizationHandler.query(filter, resultsHandler, options);
+
+            } else {
+                throw new InvalidAttributeValueException("Unsupported object class " + objectClass);
+            }
+        } catch (Exception e) {
+            throw processException(e);
         }
     }
 
@@ -302,9 +257,7 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
 
     @Override
     public void dispose() {
-        client.close();
         this.client = null;
-        this.client2 = null;
     }
 
     @Override
@@ -334,7 +287,7 @@ public class Auth0Connector implements PoolableConnector, CreateOp, UpdateDeltaO
         return new ConnectorException(e);
     }
 
-    private ConnectorException processException(Auth0Exception e) {
+    protected ConnectorException processException(Auth0Exception e) {
         if (e instanceof RateLimitException) {
             return RetryableException.wrap(e.getMessage(), e);
         }
